Move `DeliveryHelperService` to `oat\taoProctoring\model\execution\DeliveryExecution` namespace
