+++ helpers/DeliveryHelper.php
@@ -27,6 +27,7 @@ use oat\taoDeliveryRdf\model\DeliveryAssemblyService;
 use oat\taoProctoring\model\DeliveryExecutionStateService;
 use oat\taoProctoring\model\execution\DeliveryExecution;
 use oat\taoProctoring\model\execution\DeliveryExecutionManagerService;
+use oat\taoProctoring\model\execution\DeliveryHelperService;
 use oat\taoProctoring\model\monitorCache\DeliveryMonitoringService;
 use oat\taoProctoring\model\ReasonCategoryService;
 use oat\taoQtiTest\models\event\QtiTestStateChangeEvent;
@@ -35,7 +36,7 @@ use qtism\runtime\tests\AssessmentTestSessionState;
 use tao_helpers_Date as DateHelper;
 
 /**
- * @deprecated please use configurable service \oat\taoProctoring\helpers\DeliveryHelperService
+ * @deprecated please use configurable service oat\taoProctoring\model\execution\DeliveryHelperService
  *
  * This temporary helpers is a temporary way to return data to the controller.
  * This helps isolating the mock code from the real controller one.
